refactor: rename file ws-feeds and socket-middleware
