+++ src/pages/constructor-page/constructor-page.tsx
@@ -10,7 +10,7 @@ import LoadingPage from '../../components/loading-page/loading-page';
 import { useAppSelector, useAppDispatch } from '../../hooks/rtk-hooks';
 
 import { getIngredients } from '../../services/reducers/ingredients';
-import { wsConnectionSuccess, wsInit } from '../../services/reducers/wsFeeds';
+import { wsConnectionSuccess, wsInit } from '../../services/reducers/ws-feeds/ws-feeds';
 import { selectIngredientsRequest, selectIngredientsFailed } from '../../services/selectors';
 
 import type { FC } from 'react';
